fix: ascii graph width for high metric values
For high metric values the graph always goes to the new line, making it
hard to read the values.

This commit adds dynamic calculation of the ascii graph width.

Unfortunately, asciigraph library does some approximations while
writing the graph. It means the graph can still go over the specified
width. To compensate this, the graph is 0-1 chars "thinner" than it
supposed to be.